requirements.txt
APScheduler==3.6.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
decorator==5.2.1
dotenv==0.9.9
idna==3.10
pycparser==2.22
python-dotenv==1.0.1
python-telegram-bot==13.0
pytz==2025.1
requests==2.32.3
setuptools==76.0.0
six==1.17.0
tornado==6.4.2
tzlocal==5.3.1
urllib3==1.26.17
